noizu_mnesia_versioning v0.1.3 Noizu.MnesiaVersioning.Database

The Noizu.MnesiaVersioning.Database tracks state to determine which change sets have already been applied or that have been modified since initial inclusion.

Functions

__using__(opts) (macro)

Alias all the table names in the current scope and require what's needed.

create(copying \\ [])

create(Amnesia.Table.c) :: [Amnesia.Table.o]

Create the database, it calls .create on every defined table.

create!(copying \\ [])

create!(Amnesia.Table.c) :: [Amnesia.Table.o]

Create the database, it calls .create on every defined table, raises in case of error.

destroy()

destroy() :: [Amnesia.Table.o]

Destroy the database, it calls .destroy on every defined table.

destroy!()

destroy!() :: [Amnesia.Table.o]

Destroy the database, it calls .destroy on every defined table, raises in case of error.

metadata()

metadata() :: Amnesia.Metadata.t

tables()

tables() :: [atom]

List of atoms of the defined tables in the database.

wait(timeout \\ :infinity)

wait(integer | :infinity) ::
  :ok |
  {:timeout, [atom]} |
  {:error, atom}

Wait for the database to be loaded.