noizu_mnesia_versioning v0.1.3 Noizu.MnesiaVersioning.Database.ChangeSets

The Changeset table is responsbile for tracking database changset status. Schema changes are packaged into changesets that include apply and rollback methods. When using Noizu Provided helpers the system is capable of generating the appropriate rolback steps.

Types

autoincrement()

autoincrement() :: non_neg_integer

t()

t() :: %Noizu.MnesiaVersioning.Database.ChangeSets{author: String.t, changeset: String.t, date: DateTime.t, hash: String.t, key: {String.t, String.t}, note: String.t, state: :applied | :removed | :pending | :failure}

Functions

__using__(opts) (macro)

Require the needed modules to use the table effectively.

add_copy(node, type \\ :disk)

add_copy(node, Amnesia.Table.cv) :: Amnesia.Table.o

Add a copy of the table on the given node with the given mode.

add_index(attribute)

add_index(atom) :: Amnesia.Table.o

Add the index in the table for the given attribute.

at!(position)

at!(integer) :: t | nil | no_return

Read a record based on a slot, see mnesia:dirty_slot.

attributes()

attributes() :: Keyword.t

The in order keyword list of attributes passed when the table was defined.

bag?()

bag?() :: boolean

Check if the table is a bag.

clear()

clear() :: Amnesia.Table.o

Clear the content of the table.

copying(node, to)

copying(node, Amnesia.Table.cv) :: Amnesia.Table.o

Change the copying mode of the table on the given node, see mnesia:change_table_copy_type.

Modes

:disk sets :disc_copies mode

:disk! sets :disc_only_copies mode

:memory sets :ram_copies mode

count()

count() :: non_neg_integer

Get the number of records in the table.

create(copying \\ [])

create(Amnesia.Table.c) :: Amnesia.o

Create the table with the given copying mode and inherent definition.

create!(copying \\ [])

create!(Amnesia.Table.c) :: :ok | no_return

database()

database() :: module

Return the database the table belongs to.

delete(self)

delete(any | t) :: :ok | no_return

Delete the record or the given key from the table, see mnesia:delete and mnesia:delete_object.

delete(self, lock)

delete(atom | any, t | atom) :: :ok | no_return

Delete the record or the given key from the table with the given lock, see mnesia:delete and mnesia:delete_object.

Locks

:write sets a :write lock

:write! sets a :sticky_write lock

delete!(self)

delete!(t | any) :: :ok | no_return

Delete the record or the given key from the table, see mnesia:dirty_delete and mnesia:dirty_delete_object.

delete_copy(node)

delete_copy(node) :: Amnesia.Table.o

Delete a copy of the table from the given node.

delete_index(attribute)

delete_index(atom) :: Amnesia.Table.o

Delete the index in the table for the given attribute.

destroy()

destroy() :: Amnesia.Table.o

Destroy the table.

destroy!()

destroy!() :: :ok | no_return

Destroy the table, raising if an error occurs.

first(key \\ false, lock \\ :read)

first(boolean, atom) :: any | t | nil | no_return

Return the first key or record in the table, see mnesia:first.

By default it returns the record, if you want only the key pass true as first parameter.

If the table is a bag, it will return a list of records.

first!(key \\ false)

first!(boolean) :: any | t | nil | no_return

Return the first key or record in the table, see mnesia:dirty_first.

By default it returns the record, if you want only the key pass true as first parameter.

If the table is a bag, it will return a list of records.

foldl(acc, fun)

foldl(any, (t, any -> any)) :: any | no_return

Fold the whole table from the left, see mnesia:foldl.

foldr(acc, fun)

foldr(any, (t, any -> any)) :: any | no_return

Fold the whole table from the right, see mnesia:foldr.

force()

force() :: :yes | {:error, any}

Force load the table.

from_change_set(changeset, outcome)

hook_read(_, _)

hook_read!(_, _)

hook_write(_)

hook_write!(_)

id()

id() :: atom

Get the name of the id key.

info(key)

info(atom) :: any

Get information about the table, see mnesia:table_info.

key(change_sets)

key(t) :: any

Return the key of the record.

keys()

keys() :: list | no_return

Return all the keys in the table, see mnesia:all_keys.

keys!()

keys!() :: list | no_return

Return all keys in the table, see mnesia:dirty_all_keys.

last(key \\ false, lock \\ :read)

last(boolean, atom) :: any | t | nil | no_return

Return the last key or record in the table, see mnesia:last.

By default it returns the record, if you want only the key pass true as first parameter.

If the table is a bag, it will return a list of records.

last!(key \\ false)

last!(boolean) :: any | t | nil | no_return

Return the last key or record in the table, see mnesia:dirty_last.

By default it returns the record, if you want only the key pass true as first parameter.

If the table is a bag, it will return a list of records.

lock(mode)

lock(:write | :write! | :read) :: [node] | :ok | no_return

Lock the table with the given lock.

Locks

:write sets a :write lock

:write! sets a :sticky_write lock

:read sets a :read lock

majority(value)

majority(boolean) :: Amnesia.Table.o

Change the table majority.

master_nodes(nodes)

master_nodes([node]) :: :ok | {:error, any}

Set master nodes for the table, see mnesia:set_master_nodes.

match(lock \\ :read, pattern)

match(:read | :write, any) :: [t] | nil | no_return

Select records in the table using simple don't care values, see mnesia:match_object.

match!(pattern)

match!(any) :: [t] | nil | no_return

Select records in the table using simple don't care values, see mnesia:dirty_match_object.

member?(key)

member?(any) :: boolean

Check if the key is present in the table.

mode(value)

mode(:both | :read!) :: Amnesia.Table.o

Change the access of the table, see mnesia:change_table_access_mode.

Modes

:both sets read and write mode, it's the default.

:read! sets read-only mode.

move_copy(from, to)

move_copy(node, node) :: Amnesia.Table.o

Move a copy of the table from the given node to another given node.

next(self)

next(any | t) :: any | t | nil | no_return

Return the next key or record in the table, see mnesia:next.

If you're calling this function from an instance of the table (a record in it), it will get you the next record, if you're calling it directly on the module it will treat the argument as key to start from and return you the next key.

next!(self)

next!(any | t) :: any | t | nil | no_return

Return the next key or record in the table, see mnesia:dirty_next.

If you're calling this function from an instance of the table (a record in it), it will get you the next record, if you're calling it directly on the module it will treat the argument as key to start from and return you the next key.

options()

options() :: Keyword.t

The options passed when the table was defined.

ordered_set?()

ordered_set?() :: boolean

Check if the table is an ordered set.

prev(self)

prev(any | t) :: any | t | nil | no_return

Return the previous key or record in the table, see mnesia:prev.

If you're calling this function from an instance of the table (a record in it), it will get you the previous record, if you're calling it directly on the module it will treat the argument as key to start from and return you the previous key.

prev!(self)

prev!(any | t) :: any | t | nil | no_return

Return the previous key or record in the table, see mnesia:dirty_prev.

If you're calling this function from an instance of the table (a record in it), it will get you the previous record, if you're calling it directly on the module it will treat the argument as key to start from and return you the previous key.

priority(value)

priority(integer) :: Amnesia.Table.o

Change the table loading priority.

properties()

properties() :: Keyword.t

Return properties of the table.

read(key, lock \\ :read)

read(any, :read | :write | :write!) :: t | nil | no_return

Read a record from the table with the given lock, see mnesia:read.

Unlike mnesia:read this returns either the record or nil.

Locks

:write sets a :write lock

:write! sets a :sticky_write lock

:read sets a :read lock

read!(key)

read!(any) :: t | nil | no_return

Read a record from the table, see mnesia:dirty_read.

Unlike mnesia:dirty_read this returns either the record or nil.

read_at(key, position)

read_at(any, integer | atom) :: [t] | nil | no_return

Read records from the table based on a secondary index given as position, see mnesia:index_read.

read_at!(key, position)

read_at!(any, integer | atom) :: [t] | nil | no_return

Read records from the table based on a secondary index given as position, see mnesia:dirty_index_read.

select(spec)

select(any) :: Amnesia.Table.Selection.t | nil | no_return

Select records in the table using a match_spec, see mnesia:select.

select(lock_or_limit, spec)

select(integer | :read | :write, any) ::
  Amnesia.Table.Selection.t |
  nil |
  no_return

Select records in the given table using a match_spec passing a limit or a lock kind, see mnesia:select.

select(lock_or_limit, limit_or_lock, spec)

select(integer | :read | :write, integer | :read | :write, integer) ::
  Amnesia.Table.Selection.t |
  nil |
  no_return

Select records in the given table using a match_spec passing a limit and a lock kind, see mnesia:select.

select!(spec)

select!(any) :: Selection.t | nil | no_return

Select records in the table using a match_spec, see mnesia:dirty_select.

set?()

set?() :: boolean

Check if the table is a set.

stream(lock \\ :read)

stream(:read | :write | :write!) :: Amnesia.Table.Stream.t

Return an iterator to use with Enum functions.

stream!()

stream!() :: Amnesia.Table.Stream.t

Return an iterator to use with the Enum functions using dirty operations to retrieve information.

type(self)

type(t) :: :set | :ordered_set | :bag

Return the type of the table.

wait(timeout \\ :infinity)

wait(integer | :infinity) ::
  :ok |
  {:timeout, [atom]} |
  {:error, atom}

Wait for the table optionally with a timeout.

where(spec, options \\ []) (macro)

Select records in the table using an Exquisite query, see Exquisite.match/2 and mnesia:select.

Options

limit - sets the count of elements to select in every continuation

lock - sets the kind of lock to use

select - Exquisite selector spec

qualified - whether to set a name for the record or not

where!(spec, options \\ []) (macro)

Select records in the table using an Exquisite query, see Exquisite.match/2 and mnesia:dirty_select.

Options

select - Exquisite selector spec

qualified - whether to set a name for the record or not

write(self, lock \\ :write)

Write the record to the table, see mnesia:write.

Missing fields tagged as autoincrement will be incremented with the counter if nil.

write!(self)

write!(t) :: t | no_return

Write the record to the table, see mnesia:dirty_write.

Missing fields tagged as autoincrement will be incremented with the counter if nil.