noizu_mnesia_versioning v0.1.3 Noizu.MnesiaVersioning.SchemaBehaviour behaviour

This method provides information about the changesets we will be running. Currently only the change_sets() method is provided which should return a list of changesets structures to execute. In the future we will provide support for directory scanning, Similiar to how the test folders works.

Callbacks

change_sets()

change_sets() :: [Noizu.MnesiaVersioning.ChangeSet]

Return array of changesets to be applied/rolledback. @note the current logic is fairly crude. One may simply implement a module that returns an inline array of changesets.

defmodule MyApp.SchemaVersioning do
  @behaviour Noizu.MnesiaVersioning.SchemaBehaviour
  def change_sets() do
  [
    %ChangeSet{...},
    %ChangeSet{...},
  ]
  end
end

If desired, however, one could ofcourse put changesets into seperate module files and simply concatenate them together here. defmodule MyApp.SchemaVersioning do @behaviour Noizu.MnesiaVersioning.SchemaBehaviour def changesets() do # where the following each return an array of change sets. MyApp.SchemaVersioning.UserFeature.changesets() ++ MyApp.SchemaVersioning.PokerGameFeature.change_sets() end end